+++ theme/template/posts/news.php
@@ -1,0 +1,11 @@
+<section class="wrapper">
+  <div class="container">
+    <div class="row">
+      <div class="col-12 col-sm-10 offset-sm-1">
+        <main class="article">
+          <?= $post->post_content ?>
+        </main>
+      </div>
+    </div>
+  </div>
+</section>